add ingress node header for all structure definitions
